disable gzip by default
