+++ clients/algoliasearch-client-java/algoliasearch/src/main/java/com/algolia/internal/JsonSerializer.java
@@ -1,11 +1,16 @@
-package com.algolia;
+package com.algolia.internal;
 
 import com.algolia.exceptions.AlgoliaApiException;
 import com.algolia.exceptions.AlgoliaRuntimeException;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JavaType;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.*;
+
 import java.io.IOException;
+import java.util.function.Consumer;
+
+import com.fasterxml.jackson.databind.json.JsonMapper;
 import okhttp3.MediaType;
 import okhttp3.RequestBody;
 import okhttp3.Response;
@@ -17,18 +22,22 @@ import org.jetbrains.annotations.NotNull;
  * Utility class for JSON serialization and deserialization using Jackson. It provides functionality
  * to convert Java objects to their JSON representation and vice versa.
  */
-final class JsonSerializer {
+public final class JsonSerializer {
 
   private static final MediaType JSON_MEDIA_TYPE = MediaType.parse("application/json");
 
   private final ObjectMapper mapper;
+
+  public static Builder builder() {
+    return new Builder();
+  }
 
   /**
    * Initializes a new JsonSerializer instance with a given ObjectMapper.
    *
    * @param mapper The Jackson ObjectMapper to be used for JSON operations.
    */
-  public JsonSerializer(@NotNull ObjectMapper mapper) {
+  JsonSerializer(@NotNull ObjectMapper mapper) {
     this.mapper = mapper;
   }
 
@@ -109,5 +118,48 @@ final class JsonSerializer {
    */
   public JavaType getJavaType(@NotNull TypeReference<?> returnType) {
     return mapper.getTypeFactory().constructType(returnType);
+  }
+
+  public static class Builder {
+
+    /**
+     * A custom configuration for the JsonMapper builder, allowing clients to customize
+     * the built ObjectMapper beyond the default settings provided in this class.
+     */
+    private Consumer<JsonMapper.Builder> customerConfig;
+
+    /**
+     * Sets a custom configuration to be applied to the JsonMapper builder.
+     *
+     * @param config A consumer function that specifies additional configurations for the JsonMapper builder.
+     * @return The current JsonConfig instance, useful for method chaining.
+     */
+    public Builder setCustomConfig(Consumer<JsonMapper.Builder> config) {
+      this.customerConfig = config;
+      return this;
+    }
+
+    /**
+     * Builds JsonSerializer instance.
+     */
+    public JsonSerializer build() {
+      JsonMapper.Builder builder = JsonMapper
+              .builder()
+              .disable(MapperFeature.ALLOW_COERCION_OF_SCALARS)
+              .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+              .enable(JsonGenerator.Feature.AUTO_CLOSE_JSON_CONTENT)
+              .enable(DeserializationFeature.FAIL_ON_INVALID_SUBTYPE)
+              .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
+              .disable(SerializationFeature.WRITE_DATE_TIMESTAMPS_AS_NANOSECONDS)
+              .disable(DeserializationFeature.READ_DATE_TIMESTAMPS_AS_NANOSECONDS)
+              .enable(SerializationFeature.WRITE_ENUMS_USING_TO_STRING)
+              .enable(DeserializationFeature.READ_ENUMS_USING_TO_STRING)
+              .serializationInclusion(JsonInclude.Include.NON_NULL);
+      if (customerConfig != null) {
+        customerConfig.accept(builder);
+      }
+      JsonMapper build = builder.build();
+      return new JsonSerializer(build);
+    }
   }
 }
